(GH12) trying to add announce objects
